[REFACTOR][RUNTIME] Relocate nvtx.h to tvm/support/cuda and make it header-only
The NVTXScopedRange utility is a thin RAII wrapper over
nvtxRangePush/Pop with a no-op fallback when NVTX is not enabled. The
two function bodies and the conditional include of <nvtx3/nvToolsExt.h>
fit naturally inline in the header, eliminating the separate translation
unit and its TVM_RUNTIME_DLL export annotations.

- Move include/tvm/runtime/nvtx.h to include/tvm/support/cuda/nvtx.h
  under namespace tvm::support; delete src/runtime/nvtx.cc.
- Inline the constructor/destructor; gate the real-vs-stub split with
  TVM_NVTX_ENABLED in the header.
- Switch the CMake gate from a per-file COMPILE_DEFINITIONS on nvtx.cc
  to a global add_compile_definitions(TVM_NVTX_ENABLED=1) when USE_CUDA
  AND USE_NVTX, so every TU that includes the header agrees on the
  definition.
- Update the three call-site files (vm.cc, paged_kv_cache.cc,
  attn_utils.h) to the new include path and qualify NVTXScopedRange as
  support::NVTXScopedRange.
