--- src/runtime/nvtx.cc
@@ -1,42 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-#include <string>
-#ifndef TVM_NVTX_ENABLED
-#define TVM_NVTX_ENABLED 0
-#endif
-
-#if TVM_NVTX_ENABLED
-#include <nvtx3/nvToolsExt.h>
-#endif  // TVM_NVTX_ENABLED
-
-#include <tvm/runtime/nvtx.h>
-
-namespace tvm {
-namespace runtime {
-
-#if TVM_NVTX_ENABLED
-NVTXScopedRange::NVTXScopedRange(const char* name) { nvtxRangePush(name); }
-NVTXScopedRange::~NVTXScopedRange() { nvtxRangePop(); }
-#else
-NVTXScopedRange::NVTXScopedRange(const char* name) {}
-NVTXScopedRange::~NVTXScopedRange() {}
-#endif  // TVM_NVTX_ENABLED
-
-}  // namespace runtime
-}  // namespace tvm
